[#96] refactor test helpers
