+++ test/flash.go
@@ -4,6 +4,7 @@ import (
 	"net/http"
 
 	"go-google-scraper-challenge/helpers/log"
+	"go-google-scraper-challenge/lib/sessions"
 
 	"github.com/gorilla/securecookie"
 )
@@ -26,4 +27,27 @@ func FabricateCookieWithFlashes(flashes map[string]interface{}) *http.Cookie {
 	}
 
 	return &cookie
+}
+
+func GetFlashMessage(cookies []*http.Cookie) map[string][]string {
+	flashes := map[string][]string{}
+	for _, c := range cookies {
+		if c.Name == "google_scraper_session" {
+			decodedCookie := DecodeCookieString(c.Value)
+
+			if decodedCookie[sessions.FlashTypeSuccess] != nil {
+				flashes[sessions.FlashTypeSuccess] = decodedCookie[sessions.FlashTypeSuccess].([]string)
+			}
+
+			if decodedCookie[sessions.FlashTypeInfo] != nil {
+				flashes[sessions.FlashTypeInfo] = decodedCookie[sessions.FlashTypeInfo].([]string)
+			}
+
+			if decodedCookie[sessions.FlashTypeError] != nil {
+				flashes[sessions.FlashTypeError] = decodedCookie[sessions.FlashTypeError].([]string)
+			}
+		}
+	}
+
+	return flashes
 }
