+++ test/http.go
@@ -1,9 +1,11 @@
 package test
 
 import (
+	"net/http"
 	"net/http/httptest"
 
 	"github.com/gin-gonic/gin"
+	"github.com/onsi/ginkgo"
 )
 
 func CreateGinTestContext() (*gin.Context, *httptest.ResponseRecorder) {
@@ -11,4 +13,12 @@ func CreateGinTestContext() (*gin.Context, *httptest.ResponseRecorder) {
 	c, _ := gin.CreateTestContext(responseRecoder)
 
 	return c, responseRecoder
+}
+
+func GetCurrentPath(response *http.Response) string {
+	url, err := response.Location()
+	if err != nil {
+		ginkgo.Fail("Failed to get current path: " + err.Error())
+	}
+	return url.Path
 }
